api-service/requirements.txt
fastapi
uvicorn
pika
pyyaml
